movie/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Django==5.1.3
django-allauth==65.2.0
idna==3.10
# jwt==1.3.1
psycopg==3.2.3
psycopg2-binary==2.9.6
pycparser==2.22
PyJWT==2.10.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.1
urllib3==2.2.3
gunicorn==21.2.0